Add Excel download functionality and refactor routes for improved organization
